+++ web-ui/src/components/admin/permissions/sample-data.js
