+++ web-ui/src/components/admin/roles/Roles.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Roles from "./Roles";
-import { AppContextProvider } from "../../context/AppContext";
+import { AppContextProvider } from "../../../context/AppContext";
 
 const initialState = {
   state: {
